###功能 * 完成Android平台基础框架。 * 实现GTFile::Unzip解压单个文件/文件夹内所有文件。

+++ engine/lib/src/FrustumBounds.cpp
@@ -203,7 +203,7 @@ namespace Galaxy3D
             {
                 return ContainsResult::Out;
             }
-            else if(in_count == ps.size())
+            else if(in_count == (int) ps.size())
             {
                 in_plane_count++;
             }
